[fix] 글리치 다이어로그 기본 다이어로그에 합침
 ref: #226
