--- Assets/Code/Scripts/CutScene/GlitchDialog.cs
@@ -1,329 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class GlitchDialog : MonoBehaviour
-{
-    [Header("TV Global Volume 오브젝트")]
-    public GameObject tvGlobalVolume;
-    [Header("대화 말풍선")]
-    public GameObject talkPanel;
-    [Header("대화 텍스트")]
-    public TextMeshProUGUI talkText;
-    [Header("텍스트 웨이브")]
-    public bool waveText = false;
-    [Header("웨이브 세기")]
-    public float WaveAmount = 0.01f;
-    [Header("웨이브 속도")]
-    public float waveSpeed = 15f;
-    [Header("미세 흔들림")]
-    public bool shakeText = false;
-    [Header("흔들림 세기")]
-    public float shakeAmount = 0.07f;
-    [Header("흔들림 속도")]
-    public float shakeSpeed = 15f;
-    [Header("폰트 크기 연출")]
-    public float sizeUpMultiplier = 1.3f;
-    List<bool> bigCharStates = new List<bool>();
-    bool isBigMode = false;
-    [Header("| 0.1초 숨 고름, ++ 커짐, -- 원래대로")]
-    [TextArea]
-    public List<string> dialogList = new List<string>();
-
-    int currentDialogIndex = 0;
-
-    [Header("타이핑 속도")]
-    public float typingSpeed = 0.05f;
-    bool isTyping;
-
-    [HideInInspector]
-    public bool isAction;
-
-    Coroutine typingCoroutine;
-
-    public int cutscenePlayerIndex = 0; // 컷신용 플레이어 대사 인덱스
-    int cutsceneNPCIndex = 0;           // 컷신용 NPC 대사 인덱스
-
-    void Update()
-    {
-        // 테스트용 시작 키 T
-        if (!isAction &&
-            Keyboard.current != null &&
-            Keyboard.current.tKey.wasPressedThisFrame)
-        {
-            StartTestDialog();
-            return;
-        }
-
-        if (!isAction) return;
-
-        if (TimelineController.isTimelinePlaying) return;	// 컷씬 재생 중일 때는 스킵 불가
-        // 스킵 버튼
-        if (Keyboard.current != null &&
-            Keyboard.current.qKey.wasPressedThisFrame)
-        {
-            HandleEnter();
-        }
-    }
-    void StartTestDialog()
-    {
-        isAction = true;
-        currentDialogIndex = 0;
-        StartDialog(currentDialogIndex);
-    }
-
-    public void HandleEnter()
-    {
-        if (isTyping)
-        {
-            StopCoroutine(typingCoroutine);
-            ShowFullText(dialogList[currentDialogIndex]);
-            isTyping = false;
-            SetTVGlobalVolume(false);
-        }
-        else
-        {
-            NextDialog();
-        }
-    }
-
-    void ShowFullText(string text)
-    {
-        talkText.text = "";
-        bigCharStates.Clear();
-        isBigMode = false;
-
-        for (int i = 0; i < text.Length; i++)
-        {
-            if (i + 1 < text.Length && text[i] == '+' && text[i + 1] == '+')
-            {
-                isBigMode = true;
-                i++;
-                continue;
-            }
-
-            if (i + 1 < text.Length && text[i] == '-' && text[i + 1] == '-')
-            {
-                isBigMode = false;
-                i++;
-                continue;
-            }
-
-            char c = text[i];
-            if (c == '|') continue;
-
-            talkText.text += c;
-            bigCharStates.Add(isBigMode);
-        }
-
-        isTyping = false;
-    }
-
-    void NextDialog()
-    {
-        currentDialogIndex++;
-
-        if (currentDialogIndex >= dialogList.Count)
-        {
-            isAction = false;
-            talkText.text = "";
-            SetTVGlobalVolume(false);
-            StopAllCoroutines();
-            return;
-        }
-
-        StartDialog(currentDialogIndex);
-    }
-
-    public void Action()
-    {
-        isAction = !isAction;
-
-        if (isAction)
-        {
-            currentDialogIndex = 0;
-            StartDialog(currentDialogIndex);
-        }
-        else
-        {
-            StopAllCoroutines();
-        }
-    }
-    void StartDialog(int index)
-    {
-        if (typingCoroutine != null)
-            StopCoroutine(typingCoroutine);
-
-        typingCoroutine = StartCoroutine(TypeText(dialogList[index]));
-    }
-
-    // 컷신 시작
-    public void StartCutscenePlayerDialog()
-    {
-        isAction = true;
-        talkPanel.SetActive(true);
-        StartDialog(cutscenePlayerIndex);
-    }
-
-    public void StartCutsceneNPCDialog()
-    {
-        isAction = true;
-        talkPanel.SetActive(true);
-        StartDialog(cutsceneNPCIndex);
-    }
-
-
-    // 컷신 다음 대사
-    public void NextCutscenePlayerDialog()
-    {
-        cutscenePlayerIndex++;
-        StartDialog(cutscenePlayerIndex);
-    }
-
-    public void NextCutsceneNPCDialog()
-    {
-        cutsceneNPCIndex++;
-        StartDialog(cutsceneNPCIndex);
-    }
-
-
-    // 컷신 종료
-    public void EndCutscenePlayerDialog()
-    {
-        cutscenePlayerIndex++;
-        EndCutsceneDialog();
-    }
-
-    public void EndCutsceneNPCDialog()
-    {
-        cutsceneNPCIndex++;
-        EndCutsceneDialog();
-    }
-    public void EndCutsceneDialog()
-    {
-        isAction = false;
-        talkText.text = "";
-        SetTVGlobalVolume(false);
-        talkPanel.SetActive(false);
-        StopAllCoroutines();
-    }
-
-    void SetTVGlobalVolume(bool on)
-    {
-        if (tvGlobalVolume != null)
-            tvGlobalVolume.SetActive(on);
-    }
-
-    IEnumerator TypeText(string text)
-    {
-        isTyping = true;
-        SetTVGlobalVolume(true);
-        talkText.text = "";
-
-        bigCharStates.Clear();
-        isBigMode = false;
-
-        StartCoroutine(AnimateText());
-
-        for (int i = 0; i < text.Length; i++)
-        {
-            // ++
-            if (i + 1 < text.Length && text[i] == '+' && text[i + 1] == '+')
-            {
-                isBigMode = true;
-                i++;
-                continue;
-            }
-
-            // --
-            if (i + 1 < text.Length && text[i] == '-' && text[i + 1] == '-')
-            {
-                isBigMode = false;
-                i++;
-                continue;
-            }
-
-            char c = text[i];
-
-            if (c == '|')
-            {
-                yield return new WaitForSeconds(0.1f);
-                continue;
-            }
-
-            talkText.text += c;
-            bigCharStates.Add(isBigMode);
-
-            GameManager.Instance.audioManager.TextTypingSound(1f);
-            yield return new WaitForSeconds(typingSpeed);
-        }
-
-        isTyping = false;
-        SetTVGlobalVolume(false);
-    }
-
-    IEnumerator AnimateText()
-    {
-        while (isAction)
-        {
-            talkText.ForceMeshUpdate();
-            TMP_TextInfo textInfo = talkText.textInfo;
-
-            for (int i = 0; i < textInfo.characterCount; i++)
-            {
-                var charInfo = textInfo.characterInfo[i];
-                if (!charInfo.isVisible) continue;
-
-                int meshIndex = charInfo.materialReferenceIndex;
-                int vertexIndex = charInfo.vertexIndex;
-
-                Vector3[] vertices = textInfo.meshInfo[meshIndex].vertices;
-
-                Vector3 center =
-                    (vertices[vertexIndex + 0] +
-                     vertices[vertexIndex + 2]) * 0.5f;
-
-                Vector3 offset = Vector3.zero;
-
-                // 웨이브
-                if (waveText)
-                {
-                    offset.y += Mathf.Sin(Time.time * waveSpeed + i) * WaveAmount;
-                }
-
-                // 쉐이크
-                if (shakeText)
-                {
-                    offset.x += (Mathf.PerlinNoise(Time.time * shakeSpeed, i) - 0.5f) * shakeAmount;
-                    offset.y += (Mathf.PerlinNoise(i, Time.time * shakeSpeed) - 0.5f) * shakeAmount;
-                }
-
-                // 사이즈
-                float scale = 1f;
-                if (i < bigCharStates.Count && bigCharStates[i])
-                    scale = sizeUpMultiplier;
-
-                for (int v = 0; v < 4; v++)
-                {
-                    Vector3 pos = vertices[vertexIndex + v];
-                    pos = (pos - center) * scale + center;
-                    pos += offset;
-                    vertices[vertexIndex + v] = pos;
-                }
-            }
-
-            for (int i = 0; i < textInfo.meshInfo.Length; i++)
-            {
-                textInfo.meshInfo[i].mesh.vertices =
-                    textInfo.meshInfo[i].vertices;
-                talkText.UpdateGeometry(
-                    textInfo.meshInfo[i].mesh, i);
-            }
-
-            yield return null;
-        }
-    }
-
-}
