--- Assets/Code/Scripts/CutScene/GlitchDialog.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bef67e65d8ce2e240a9d576c5fb699cd
